[ISSUE #4423] Fix the meta-file error in the meta module
