+++ build.gradle
@@ -279,10 +279,9 @@ subprojects {
         Set<String> rootProject = ["eventmesh-admin",
                                    "eventmesh-admin-rocketmq",
                                    "eventmesh-common",
-                                   "eventmesh-connector-api",
                                    "eventmesh-storage-api",
                                    "eventmesh-metrics-api",
-                                   "eventmesh-registry-api",
+                                   "eventmesh-meta-api",
                                    "eventmesh-trace-api",
                                    "eventmesh-runtime",
                                    "eventmesh-security-api",
