--- eventmesh-meta/eventmesh-meta-consul/src/main/java/org/apache/eventmesh/registry/consul/config/ConsulTLSConfig.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.eventmesh.registry.consul.config;
-
-import org.apache.eventmesh.common.config.Config;
-import org.apache.eventmesh.common.config.ConfigFiled;
-import org.apache.eventmesh.common.config.convert.converter.EnumConverter;
-
-import com.ecwid.consul.transport.TLSConfig.KeyStoreInstanceType;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@Config(prefix = "eventMesh.registry.consul.tls")
-public class ConsulTLSConfig {
-
-    @ConfigFiled(field = "keyStoreInstanceType", converter = EnumConverter.class)
-    private KeyStoreInstanceType keyStoreInstanceType;
-
-    @ConfigFiled(field = "certificatePath")
-    private String certificatePath;
-
-    @ConfigFiled(field = "certificatePassword")
-    private String certificatePassword;
-
-    @ConfigFiled(field = "keyStorePath")
-    private String keyStorePath;
-
-    @ConfigFiled(field = "keyStorePassword")
-    private String keyStorePassword;
-
-}
