--- eventmesh-meta/eventmesh-meta-etcd/src/main/java/org/apache/eventmesh/registry/etcd/factory/EtcdStreamObserver.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.eventmesh.registry.etcd.factory;
-
-import io.etcd.jetcd.lease.LeaseKeepAliveResponse;
-import io.grpc.stub.StreamObserver;
-
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class EtcdStreamObserver implements StreamObserver<LeaseKeepAliveResponse> {
-
-    private EtcdLeaseId etcdLeaseId;
-
-    @Override
-    public void onNext(LeaseKeepAliveResponse value) {
-    }
-
-    @Override
-    public void onError(Throwable t) {
-        log.debug("EtcdStreamObserver lease renewal Exception", t);
-    }
-
-    @Override
-    public void onCompleted() {
-        log.info("EtcdStreamObserver completed");
-    }
-
-    public void setEtcdLeaseId(EtcdLeaseId etcdLeaseId) {
-        this.etcdLeaseId = etcdLeaseId;
-    }
-}
